+++ R/vapour-package.R
@@ -82,3 +82,13 @@ NULL
 NULL
 
 
+#' Example WKT coordinate reference system
+#'
+#' A Lambert Azimuthal Equal Area Well-Known-Text string for a region
+#' centred on Tasmania.
+#'
+#' Created from '+proj=laea +lon_0=147 +lat_0=-42 +datum=WGS84'.
+#' For use in a future warping example.
+#' @docType data
+#' @name tas_wkt
+NULL