+++ NEWS.md
@@ -1,3 +1,31 @@
+# vapour 0.4.0
+
+## BREAKING CHANGES
+
+* The function `vapour_read_raster()` now returns a list of (one) numeric vector. This change 
+ is so a future version can return more than one band for this and related functions. 
+ 
+## NEW FEATURES
+
+* New data set `tas_wkt` a Well-Known-Text string to simplify an example
+ in the documentation. 
+ 
+* The function `vapour_read_raster()` now returns a list with a single vector of
+ values from the requested band. In future this may return values from multiple
+ bands. The new function `vapour_warp_raster()` uses the same scheme.
+
+* Fixed static linking for MacOS build, thanks(!!) to Jeroen Ooms for guidance 
+ and thanks to CRAN for reporting the issue. 
+ 
+* New function `vapour_srs_wkt()` to convert PROJ4 strings to Well-Known Text. 
+
+* New *internal-only* function `vapour_warp_raster()` to return a warped version from
+ a raster source. Please use with caution, it's still flaky for certain cases and will
+ crash your session (setting projection on NetCDF 0,360 is one example but not fleshed
+ out yet - see hypertidy/lazyraster for in-dev usage examples.)
+
+* New internal function `warp_memory_cpp()` to warp small in memory rasters. 
+
 # vapour 0.3.0
 
 * Updates to configure to remove deprecated R CMD config CPP (thanks to CRAN).